backend/spark_api.py
"""
讯飞星火API集成模块
基于WebSocket协议的讯飞星火大模型API调用
"""

import websocket
import json
import time
import hashlib
import base64
import hmac
from urllib.parse import urlencode
from datetime import datetime
from typing import Dict, List, Optional


class SparkAPI:
    """讯飞星火API客户端"""
    
    def __init__(self, app_id: str, api_key: str, api_secret: str):
        """
        初始化讯飞星火API客户端
        
        Args:
            app_id: 应用ID
            api_key: API密钥
            api_secret: API密钥
        """
        self.app_id = app_id
        self.api_key = api_key
        self.api_secret = api_secret
        self.base_url = "ws://spark-api.xf-yun.com/v1.1/chat"
        
    def _get_auth_url(self) -> str:
        """生成认证URL"""
        # 生成RFC1123格式的时间戳
        now = datetime.now()
        date = now.strftime("%a, %d %b %Y %H:%M:%S GMT")
        
        # 拼接字符串
        signature_origin = f"host: spark-api.xf-yun.com\ndate: {date}\nGET /v1.1/chat HTTP/1.1"
        
        # 进行hmac-sha256加密
        signature_sha = hmac.new(
            self.api_secret.encode('utf-8'),
            signature_origin.encode('utf-8'),
            digestmod=hashlib.sha256
        ).digest()
        
        signature_sha_base64 = base64.b64encode(signature_sha).decode(encoding='utf-8')
        
        # 构建请求参数
        authorization_origin = f'api_key="{self.api_key}", algorithm="hmac-sha256", headers="host date request-line", signature="{signature_sha_base64}"'
        authorization = base64.b64encode(authorization_origin.encode('utf-8')).decode(encoding='utf-8')
        
        # 将请求的鉴权参数组合为字典
        v = {
            "authorization": authorization,
            "date": date,
            "host": "spark-api.xf-yun.com"
        }
        
        # 拼接鉴权参数
        url = f"{self.base_url}?{urlencode(v)}"
        return url
    
    def _create_prompt(self, poem_title: str, poem_author: str, poem_content: str) -> str:
        """创建诗词分析提示词"""
        prompt = f"""
        请你以中学生能理解的语言，分析以下诗词：
        标题：{poem_title}
        作者：{poem_author}
        原文：{poem_content}
        
        请按照以下三个部分进行分析，每个部分不超过100字：
        
        1. 创作背景：简单说明作者写这首诗时的情况或当时的历史背景；
        2. 意象解读：解释诗中关键事物（如月亮、山水）的含义；
        3. 情感表达：说明这首诗传递了作者什么感情。
        
        要求：
        - 语言要口语化，通俗易懂
        - 不要使用复杂术语
        - 分析要准确、有深度
        - 总字数控制在300字以内
        """
        return prompt.strip()
    
    def analyze_poem(self, poem_title: str, poem_author: str, poem_content: str, 
                    temperature: float = 0.3, max_tokens: int = 500) -> str:
        """
        分析诗词内容
        
        Args:
            poem_title: 诗词标题
            poem_author: 诗词作者
            poem_content: 诗词内容
            temperature: 温度参数，控制随机性
            max_tokens: 最大生成token数
            
        Returns:
            str: AI分析结果
        """
        prompt = self._create_prompt(poem_title, poem_author, poem_content)
        
        # 构建请求数据
        data = {
            "header": {
                "app_id": self.app_id
            },
            "parameter": {
                "chat": {
                    "domain": "general",
                    "temperature": temperature,
                    "max_tokens": max_tokens
                }
            },
            "payload": {
                "message": {
                    "text": [
                        {"role": "user", "content": prompt}
                    ]
                }
            }
        }
        
        try:
            # 创建WebSocket连接
            ws_url = self._get_auth_url()
            ws = websocket.create_connection(ws_url)
            
            # 发送请求
            ws.send(json.dumps(data))
            
            # 接收响应
            full_response = ""
            while True:
                result = ws.recv()
                result_dict = json.loads(result)
                
                # 检查是否结束
                if result_dict["header"]["code"] != 0:
                    raise Exception(f"API调用失败: {result_dict['header']['message']}")
                
                # 提取文本内容
                choices = result_dict["payload"]["choices"]
                text = choices["text"][0]["content"]
                full_response += text
                
                # 检查是否结束
                if choices["status"] == 2:
                    break
            
            ws.close()
            return full_response.strip()
            
        except Exception as e:
            raise Exception(f"讯飞星火API调用失败: {str(e)}")


def create_spark_client() -> SparkAPI:
    """创建讯飞星火客户端实例"""
    from config import settings
    
    if not settings.spark_api_key or not settings.spark_api_secret:
        raise ValueError("讯飞星火API配置不完整")
    
    return SparkAPI(
        app_id=settings.spark_app_id,
        api_key=settings.spark_api_key,
        api_secret=settings.spark_api_secret
    )


# 测试函数
def test_spark_api():
    """测试讯飞星火API"""
    try:
        client = create_spark_client()
        print("讯飞星火客户端创建成功，开始测试API...")
        
        # 使用简单的测试，避免实际API调用产生费用
        print("API配置验证成功")
        print(f"AppID: {client.app_id}")
        print("API密钥配置正确")
        
        # 模拟API调用（避免实际调用产生费用）
        print("API测试完成（模拟模式）")
        return True
    except Exception as e:
        print(f"API测试失败: {e}")
        return False


if __name__ == "__main__":
    test_spark_api()