backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.11.0
redis==5.0.1
websocket-client==1.6.4
python-multipart==0.0.6